requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
tabulate>=0.9.0
python-dateutil>=2.8
pytz>=2023.3
alpaca-trade-api>=3.1.1
python-dotenv>=1.0.0
